asd
Added Rating system

+++ AndroidBackk/api/models.py
@@ -30,7 +30,7 @@ class Product(models.Model):
 class Rating(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     product = models.ForeignKey(Product, on_delete=models.CASCADE, related_name='rating')
-    rating = models.IntegerField(default=0)
+    rating = models.FloatField(default=0)
 
     def __str__(self):
         return f'{self.product.name}: {self.rating}'
